+++ lib/blockchain/validator.ex
@@ -1,6 +1,7 @@
 defmodule UltraDark.Validator do
   alias UltraDark.Blockchain.Block, as: Block
   alias UltraDark.Transaction, as: Transaction
+  alias UltraDark.Utilities, as: Utilities
 
   @doc """
     A block is considered valid if the index is greater than the index of the previous block,
@@ -24,23 +25,8 @@ defmodule UltraDark.Validator do
   defp valid_prev_hash(prev_hash, last_block_hash) when prev_hash == last_block_hash, do: :ok
 
   defp valid_hash(%{index: index, previous_hash: previous_hash, timestamp: timestamp, nonce: nonce, hash: hash}) do
-    if Block.calculate_hash([Integer.to_string(index), previous_hash, timestamp, Integer.to_string(nonce)]) == hash, do: :ok
+    if Utilities.sha_base16([Integer.to_string(index), previous_hash, timestamp, Integer.to_string(nonce)]) == hash, do: :ok
   end
-
-
-  @doc """
-    The merkle root lets us represent a large dataset using only one string. We can be confident that
-    if any of the data changes, the merkle root will be different, which invalidates the dataset
-  """
-  def calculate_merkle_root(list) do
-    list
-    |> Enum.chunk_every(2)
-    |> Enum.map(fn (h) -> :crypto.hash(:sha256, h) |> Base.encode16 end)
-    |> calculate_merkle_root(true)
-  end
-
-  def calculate_merkle_root(list, true) when length(list) == 1, do: List.first(list)
-  def calculate_merkle_root(list, true), do: calculate_merkle_root(list)
 
   def valid_coinbase?(%{transactions: transactions, index: block_index}) do
     coinbase = List.first(transactions)
