+++ lib/ultradark.ex
@@ -17,7 +17,7 @@ defmodule UltraDark do
     |> Block.initialize
     |> Block.mine
 
-    IO.puts "\e[34mBlock hash calculated:\e[0m #{block.hash}, using nonce: #{block.nonce}"
+    IO.puts "\e[34mBlock hash at index #{block.index} calculated:\e[0m #{block.hash}, using nonce: #{block.nonce}"
 
     if Validator.is_block_valid?(block, chain) do
       main(Blockchain.add_block(chain, block))
